Testing: makefile.GF now compiles and runs felixrefine on tinis.Unlike previous commit.
Why:
Problems with recognising the tinis shell in makefile.GF (unsure why?). Updated the makefile so that it now does recognise the tinis shell and all shells Richard and I currently use. I've tested compilation on tinis and CSC. only runs on tinis due to read/write bit issue

Implementation:
Current makefile.GF is in an if-else structure. To recognise the shell the command  '$(shell (lsb_release -si))' is used (line54).
This doesn't seem to be recognised as a valid function in the tinis shell. I will find out why at some point in the future.
But for now I have taken the environmental variable 'MODULEPATH' and found the word 'tinis' in it. I am hoping this is a default path and never changes (or at least until I find out what is going wrong with the shell identification function)

TODO:
Figure out how to implement a byte/bit conversion to swap between tinis/CSC