Testing: Bypassed Read in of Experimental images in sim mode
Why:
There is no need to read in any experimental images when you simulate the
images

Implementation:
If statement in felixrefine.f90 over subroutine ReadExperimentalImages.
The output has changed and is confusing me for Refine. IReflectOut is not
working.

TODO:
test on tinis
test refine to ensure it still works