Started introducing a grid simulation mode for Atomic Refinement
Why:
I would like to see what the refinement space looks like so that I can choose
an appropriate algorithm

Implementation:
Started in Read_files_mod ISimFLAG=2 is now grid simulation - will be represented
as a 'T' in the input file for SRefineMode. The format is (a:b,c:d) the first
number (a & c) is the same Atom recogniser, the number after the colon (b & d)
will be the number of points in the range 0->1 (we are in fractional
coordinates) so if b = 1 we will only have three simulations, when the atom is at zero, One and 0.5. For b=2 wr will have 4 simulations etc.

TODO:
Finish the reading in of the input file
move on to felixrefine where a loop of all the simulations will be put in place
we only need to call SimulateAndFit and modify it to accept the ISimFlag.EQ.2