Finished read in implementation of new grid sim format
Why:
The read in part looks a bit of a mess - took a while to understand what
was going on. The section needs to be summerised (commented) properly

Implementation:
Added sections to DetermineRefineableAtomicSites to give an extra number
separated by a colon. In future I will hopefully be able to introduce reals to
signify the range of the read in

TODO:
Test the read in then implement the grid simulation mode in felixrefine