Debugging Process of read in  - Simplified SAtomicSites
Why:
SAtomicSites read in the full line of text from the input file and all the
character positions were just referenced from the full line.
Absolute nightmare to follow! I simply cut down the read in line to the bracketed
numbers.I must admit semi-frustrating

Implementation:
SAtomicSites now read in as SFullAtomicSites which is the full line in the input
file, SAtomicSites is now only the bracketed characters. The Character positions
now make much more sense. With '(' referencing '1' instead of '28'

TODO:
There is a bug in reading the grid values. The next step is to find and correct
I can then finally implement the hard-coded grid sims