Changed the cif to read in characters not numbers to ensure precision errors are at a minimum
Why:
The ciftoolbox reads in single precision numbers. We then convert that into
Double precision. It is therefore easier to read the cif as characters and then
straight into a double precision variable. Thus removing the need for single
precision entirely

Implementation:
New subroutine called RemoveBrackets - if there is a bracket in the number to
signify error we remove it

TODO:
The big influence bug still exists in Cave version of felix. It happens before
the Ug matrix in Cvgij - just need to find it