Introduced structure.Ug
Why:
The Ug matrix gives a tremendous amount of information when things go wrong.
All the bugs up until now have been somewhat noticable in the Ug matrix
Therefore it is essential that we print it

Implementation:
New subroutine in WriteIterationOutput - WriteIterationUg
A bug in writing out the Ugs to the screen has been rectified. This causes havoc
on tinis

TODO:
Create new branch to collate all of the changes I have made to merge into felix